[Chore/#9] cornerRounding Util 분리
