+++ TheGreatestYourManito-iOS/TheGreatestYourManito-iOS/Global/Extension/View+.swift
@@ -1,0 +1,14 @@
+//
+//  View+.swift
+//  TheGreatestYourManito-iOS
+//
+//  Created by 박민서 on 10/28/24.
+//
+
+import SwiftUI
+
+extension View {
+    func cornerRadius(_ radius: CGFloat, corners: UIRectCorner) -> some View {
+        clipShape(RoundedCorner(radius: radius, corners: corners))
+    }
+}
