+++ TheGreatestYourManito-iOS/TheGreatestYourManito-iOS/Global/Utility/roundCorner.swift
@@ -1,0 +1,20 @@
+//
+//  roundCorner.swift
+//  TheGreatestYourManito-iOS
+//
+//  Created by 박민서 on 10/28/24.
+//
+
+import SwiftUI
+
+struct RoundedCorner: Shape {
+    var radius: CGFloat = 10.0
+    var corners: UIRectCorner = .allCorners
+    
+    func path(in rect: CGRect) -> Path {
+        let path = UIBezierPath(roundedRect: rect,
+                                byRoundingCorners: corners,
+                                cornerRadii: CGSize(width: radius, height: radius))
+        return Path(path.cgPath)
+    }
+}
