+++ TheGreatestYourManito-iOS/TheGreatestYourManito-iOS/Presentation/Participate/PlayManitto/PlayManittoView.swift
@@ -24,21 +24,3 @@ struct PlayManittoView: View {
         .ignoresSafeArea(.container, edges: .bottom)
     }
 }
-
-extension View {
-    func cornerRadius(_ radius: CGFloat, corners: UIRectCorner) -> some View {
-        clipShape(RoundedCorner(radius: radius, corners: corners))
-    }
-}
-
-struct RoundedCorner: Shape {
-    var radius: CGFloat = 10.0
-    var corners: UIRectCorner = .allCorners
-    
-    func path(in rect: CGRect) -> Path {
-        let path = UIBezierPath(roundedRect: rect,
-                                byRoundingCorners: corners,
-                                cornerRadii: CGSize(width: radius, height: radius))
-        return Path(path.cgPath)
-    }
-}
